Add new methods try permutations

+++ SortingNetworks/Extensions/BitArrayExtenstions.cs
@@ -5,12 +5,12 @@ namespace SortingNetworks.Extensions
 
     public static class BitArrayExtenstions
     {
-        public static ushort SetCount(this BitArray arr)
+        public static int SetCount(this BitArray arr)
         {
-            var newValue = new uint[1];
+            var newValue = new int[1];
             arr.CopyTo(newValue, 0);
 
-            return (ushort)System.Runtime.Intrinsics.X86.Popcnt.PopCount(newValue[0]);
+            return (int)System.Runtime.Intrinsics.X86.Popcnt.PopCount((uint) newValue[0]);
         }
     }
 }
