+++ SortingNetworksTests/ComparatorNetworkTests.cs
@@ -12,7 +12,7 @@ namespace SortingNetworksTests
         public void IsSortingNetwork_WithSize2And1Comparator_ReturnsTrue()
         {
             // Arrange, Act
-            ushort size = 2;
+            int size = 2;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[1] { new Comparator(0, 1) };
             var s1 = new ComparatorNetwork(comparators);
@@ -25,7 +25,7 @@ namespace SortingNetworksTests
         public void IsSortingNetwork_WithSize3And3Comparators_ReturnsTrue()
         {
             // Arrange, Act
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[3] { new Comparator(0, 2), new Comparator(0, 1), new Comparator(1, 2) };
             var s1 = new ComparatorNetwork(comparators);
@@ -38,7 +38,7 @@ namespace SortingNetworksTests
         public void IsSortingNetwork_WithSize3And3Comparators_ReturnsFalse()
         {
             // Arrange, Act
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[3] { new Comparator(1, 2), new Comparator(0, 1), new Comparator(0, 2) };
             var s1 = new ComparatorNetwork(comparators);
@@ -51,7 +51,7 @@ namespace SortingNetworksTests
         public void IsSortingNetwork_WithSize3And2Comparators_ReturnsFalse()
         {
             // Arrange, Act
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[2] { new Comparator(0, 2), new Comparator(0, 1) };
             var s1 = new ComparatorNetwork(comparators);
@@ -64,7 +64,7 @@ namespace SortingNetworksTests
         public void IsSortingNetwork_WithSize3And1Comparator_ReturnsFalse()
         {
             // Arrange, Act
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[1] { new Comparator(0, 1) };
             var s1 = new ComparatorNetwork(comparators);
@@ -77,7 +77,7 @@ namespace SortingNetworksTests
         public void CloneWithNewComparator_When0Comparator_Contains1Comparator()
         {
             // Arrange
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[0];
             var s1 = new ComparatorNetwork(comparators);
@@ -95,7 +95,7 @@ namespace SortingNetworksTests
         public void CloneWithNewComparator_When1Comparator_Contains2Comparators()
         {
             // Arrange
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var comparators = new Comparator[1] { new Comparator(0, 1) };
             var s1 = new ComparatorNetwork(comparators);
@@ -115,7 +115,7 @@ namespace SortingNetworksTests
         public void IsRedundant_WhenAreRedundant_IsMarkedIsTrue()
         {
             // Arrange
-            ushort size = 2;
+            int size = 2;
             IComparatorNetwork.Inputs = size;
             var c1 = new Comparator[1] { new Comparator(0, 1) };
             var c2 = new Comparator[1] { new Comparator(0, 1) };
@@ -133,7 +133,7 @@ namespace SortingNetworksTests
         public void IsRedundant_WhenAreNotRedundant_IsMarkedIsFalse()
         {
             // Arrange
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var c1 = new Comparator[1] { new Comparator(0, 1) };
             var c2 = new Comparator[1] { new Comparator(0, 2) };
@@ -151,21 +151,21 @@ namespace SortingNetworksTests
         public void Output_When3Inputs1Comparator_HasExpectedResult()
         {
             // Arrange, Act
-            ushort size = 3;
+            int size = 3;
             IComparatorNetwork.Inputs = size;
             var c1 = new Comparator[1] { new Comparator(0, 1) };
             var n = new ComparatorNetwork(c1);
 
             // Assert
             //Assert.AreEqual(4, n.Outputs.Count);
-            //Assert.IsTrue(n.Outputs.SetEquals(new HashSet<ushort> { 1, 4, 5, 6 }));
+            //Assert.IsTrue(n.Outputs.SetEquals(new HashSet<int> { 1, 4, 5, 6 }));
         }
 
         [TestMethod]
         public void SubsumePositiveTest()
         {
             // Arrange, Act
-            ushort size = 4;
+            int size = 4;
             IComparatorNetwork.Inputs = size;
             var c1 = new Comparator[3] { new Comparator(0, 1), new Comparator(0, 3), new Comparator(1,2)  };
             var c2 = new Comparator[3] { new Comparator(0, 1), new Comparator(0, 3), new Comparator(2,3) };
