+++ SortingNetworksTests/SortingNetworksGeneratorTests.cs
@@ -17,7 +17,7 @@
         public void Generate_WithRedundantComparator_NotAddNewNet()
         {
             // Arrange
-            const ushort Size = 2;
+            const int Size = 2;
             var comparators = new Comparator[] { new Comparator(0, 1), new Comparator(0, 1) };
             var generator = new Generator();
             IComparatorNetwork.Inputs = Size;
@@ -34,7 +34,7 @@
         public void Generate_WithNonRedundantComparator_AddsNewNets()
         {
             // Arrange
-            const ushort Size = 3;
+            const int Size = 3;
             var c1 = new Comparator[1] { new Comparator(0, 1) };
             var c2 = new Comparator[2] { new Comparator(0, 2), new Comparator(1, 2) };
             var generator = new Generator();
